refine(data): refactor temperature and interval fields from Float to Int
This commit updates the data models and UI components to use integer values for temperature, interval numbers, and Rate of Rise (ROR) calculations. This simplification improves data consistency and aligns with the expected precision for roasting logs.

### Key Changes:

**Data Model & Logic:**
- **Type Migration:** Changed `temperature` and `intervalNumber` from `Float` to `Int` in `IntervalData`, `RoastingEvent`, and `ChartDataPoint`.
- **ROR Calculation:** Updated `RoastingRunViewModel` to calculate ROR as an integer difference.
- **Database Version:** Incremented `RoastDatabase` version to `4` to reflect changes in the underlying data structure.
- **ViewModel Updates:** Refactored event handling logic (`addBurnerEvent`, `addAirFlowEvent`, etc.) and event marking functions (`markTurnPoint`, `markYellowing`, etc.) to accept `Int` parameters.

**UI & Visualization:**
- **Chart Logic:** Simplified `RoastingChart` filtering and mapping logic. The chart now handles `Int` based interval numbers while still converting to `Float` internally for the `MPAndroidChart` `Entry` API.
- **Input Handling:** Updated `RoastingFormScreen` and `RoastingRunScreen` to use `toIntOrNull()` for parsing user input in temperature and event dialogs.
- **PDF Export:** Updated `PdfExportManager` to support the new integer-based `startTemperature` and `ChartDataPoint` structure.

**Refinement:**
- Removed redundant float-to-int conversions in chart labels and table row filters.
- Standardized default values (e.g., `0` instead of `0f`) across view models and screen components.

+++ app/src/main/java/com/indie/roastlog/ui/model/RoastingModels.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class IntervalData(
     val intervalNumber: Int,
-    val temperature: Float,
+    val temperature: Int,
     val airFlowPower: String = "",
     val rpmDrum: String = "",
     val burnerPower: String = ""
@@ -13,7 +13,7 @@ data class IntervalData(
 
 @Serializable
 data class RoastingEvent(
-    val temperature: Float,
+    val temperature: Int,
     val seconds: Int
 ) {
     val time: String
